Tidy code and add comprehensive test for FormatConverter
Former-commit-id: dc66659dd42aed28f6a7c7456e50d3f4f7f829e6

+++ record_classification/src/main/java/uk/ac/standrews/cs/digitising_scotland/parser/datastructures/ListAccuracyMetrics.java
@@ -161,7 +161,7 @@ public class ListAccuracyMetrics {
                 }
             }
 
-            propGoldPredicted += (count / goldStandardTriples.size());
+            propGoldPredicted += (count / (double) goldStandardTriples.size());
         }
         return propGoldPredicted / bucket.size();
     }
